% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.all.lmers.R
\name{run.all.lmers}
\alias{run.all.lmers}
\title{Function that will run a regression for your variable of interest across all genes/metabolites/CpGs, in parallel.  Adapted from Vamsee (github.com/vkp3/pillalamarRi)}
\usage{
run.all.lmers(
  tx_expr,
  cov,
  rcov,
  gene.ids,
  SCORE,
  omit.outlier = T,
  num.cores = 10
)
}
\arguments{
\item{tx_expr}{Expression matrix in form: \link{genes x samples}. Will be converted to a list(!) of gene-expr vectors (if not input as list)}

\item{cov}{Regression covariates (fixed effects) \link{cov x samples}}

\item{rcov}{Regression covariates to be treated as random effects}

\item{gene.ids}{Character vector of gene IDs, corresponding to rows in \code{tx_expr} \link{genes x samples}}

\item{SCORE}{Main covariate to be permuted (not included in \code{cov})}

\item{omit.outlier}{Whether or not you want to omit gene expression outliers}

\item{num.cores}{The number of cores you would like to use}
}
\value{
All regression coefficients for lmer(gene expression ~ SCORE + cov + (1|rcov)) for all genes
}
\description{
Function that will run a regression for your variable of interest across all genes/metabolites/CpGs, in parallel.  Adapted from Vamsee (github.com/vkp3/pillalamarRi)
}
\note{

}
\examples{
tx_expr <- w.mt[,which(colnames(w.mt) \%in\% make.names(metablist))]
cov <- w.mt[,which(colnames(w.mt) \%in\% c('sex', 'age1c', 'PC1', 'PC2', 'PC3', 'bmi', 'Exam'))]
rcov <- w.mt[,which(colnames(w.mt) \%in\% c('idno'))]
SCORE <- w.mt[,which(colnames(w.mt) == "dpcrAdjMetric")]
lm_res.sort <- run.all.lmers(tx_expr, cov, rcov, colnames(tx_expr), SCORE, omit.outlier = T, num.cores = 10)

}
